fix: move BH header to public /include/
