+++ include/igraph_barnes_hut.h
